+++ back/src/main/java/com/seafood/back/entity/ProductEntity.java
@@ -12,29 +12,6 @@ import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
-// @Entity
-// @Table(name="product_list")
-// @Getter
-// @Setter
-
-// public class ProductEntity {
-//     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-//     @Column(name = "product_id")
-//     private Integer productId;
-
-//     @Column(name = "product_name")
-//     private String productName;
-
-//     @Column(name = "product_price")
-//     private Integer productPrice;
-
-//     @Column(name = "product_discount")
-//     private Integer productDiscount;
-
-//     @Column(name = "product_img")
-//     private String productImgPath;
-// }
-
 
 @Entity
 @Table(name = "products")
@@ -44,7 +21,7 @@ public class ProductEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "product_id")
-    private Integer productId;
+    private Long productId;
 
     @Column(name = "category")
     private Integer category;
