+++ back/src/main/java/com/seafood/back/respository/ProductRepository.java
@@ -8,12 +8,12 @@ import org.springframework.stereotype.Repository;
 import com.seafood.back.entity.ProductEntity;
 
 @Repository
-public interface ProductRepository extends JpaRepository<ProductEntity, Integer>{
+public interface ProductRepository extends JpaRepository<ProductEntity, Long>{
     List<ProductEntity> findTop10ByOrderByProductIdDesc();
     List<ProductEntity> findByCategory(Long category);
     List<ProductEntity> findByNameContaining(String query);
-    ProductEntity findByProductId(Integer productId);
-    List<ProductEntity> findAllByProductIdIn(List<Integer> productIds);
+    ProductEntity findByProductId(Long productId);
+    List<ProductEntity> findAllByProductIdIn(List<Long> productIds);
     List<ProductEntity> findByRecommended(boolean b);
