+++ back/src/main/java/com/seafood/back/respository/ReviewImageRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.seafood.back.entity.ReviewImageEntity;
 
-public interface ReviewImageRepository extends JpaRepository<ReviewImageEntity, Integer> {
+public interface ReviewImageRepository extends JpaRepository<ReviewImageEntity, Long> {
 
-    List<ReviewImageEntity> findByReviewId(int reviewId);
+    List<ReviewImageEntity> findByReviewId(Long reviewId);
 
     
 }
